requirements.txt
annotated-types==0.7.0
anthropic==0.41.0
anyio==4.7.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
colorama==0.4.6
Deprecated==1.2.15
distro==1.9.0
eval_type_backport==0.2.0
executing==2.1.0
google-auth==2.37.0
googleapis-common-protos==1.66.0
griffe==1.5.1
groq==0.13.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
jiter==0.8.2
json_repair==0.31.0
jsonpath-python==1.0.6
logfire-api==2.7.1
markdown-it-py==3.0.0
mdurl==0.1.2
mistralai==1.2.5
mypy-extensions==1.0.0
openai==1.57.4
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
packaging==24.2
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic-ai==0.0.13
pydantic-ai-slim==0.0.13
pydantic_core==2.27.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
rsa==4.9
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
wrapt==1.17.0
zipp==3.21.0
